chore: add new theming infrastructure
