+++ src/components/Common/FieldCaption/FieldCaption.module.scss
@@ -8,7 +8,6 @@
 }
 
 .optionalLabel {
-  content: var(--g-optionalLabel);
   font-size: smaller;
   color: var(--g-input-placeholderColor);
 }
