+++ src/contexts/ThemeProvider/ThemeProvider.tsx
@@ -3,103 +3,18 @@ import { useEffect, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { createTheme } from './createTheme'
 import { ThemeContext } from './useTheme'
+import { gustoSDKTheme, type GustoSDKTheme } from './theme'
 import type { GTheme } from '@/types/GTheme'
 import '@/styles/sdk.scss'
 import type { DeepPartial } from '@/types/Helpers'
-import { getRootFontSize, toRem } from '@/helpers/rem'
+
+// Create a union type that allows both the existing GTheme structure and the new flat structure
+type EnhancedTheme = DeepPartial<GTheme> & Partial<GustoSDKTheme>
 
 export interface ThemeProviderProps {
-  theme?: DeepPartial<GTheme>
+  theme?: EnhancedTheme
   children?: React.ReactNode
 }
-
-const colors = {
-  gray: {
-    100: '#FFFFFF',
-    200: '#FBFAFA',
-    300: '#F4F4F3',
-    400: '#EAEAEA',
-    500: '#DCDCDC',
-    600: '#BABABC',
-    700: '#919197',
-    800: '#6C6C72',
-    900: '#525257',
-    1000: '#1C1C1C',
-  },
-  error: {
-    100: '#FFF7F5',
-    500: '#D5351F',
-    800: '#B41D08',
-  },
-  warning: {
-    100: '#FFFAF2',
-    500: '#E9B550',
-    700: '#B88023',
-    800: '#B88023',
-  },
-  success: {
-    100: '#F3FAFB',
-    400: '#2BABAD',
-    500: '#0A8080',
-    800: '#005961',
-  },
-  info: {
-    100: '#F3FAFB',
-    400: '#2BABAD',
-    500: '#0A8080',
-    800: '#005961',
-  },
-  orange: {
-    800: '#CA464A',
-  },
-}
-
-const defaultTheme = {
-  // Colors
-  colorBody: colors.gray[100],
-  colorBodyContent: colors.gray[1000],
-  colorPrimary: colors.gray[1000],
-  colorPrimaryContent: colors.gray[100],
-  colorSecondary: colors.gray[100],
-  colorSecondaryContent: colors.gray[1000],
-  colorInfo: colors.info[800],
-  colorInfoContent: colors.info[100],
-  colorWarning: colors.warning[800],
-  colorWarningContent: colors.warning[100],
-  colorError: colors.error[800],
-  colorErrorContent: colors.error[100],
-  colorSuccess: colors.success[800],
-  colorSuccessContent: colors.success[100],
-  // Radius
-  radius: '6px',
-  // Font
-  fontSizeRoot: getRootFontSize(),
-  fontFamily: 'Geist',
-  fontLineHeight: '1.5rem',
-  fontSizeSmall: toRem(14),
-  fontSizeRegular: toRem(16),
-  fontSizeMedium: toRem(18),
-  fontSizeHeading1: toRem(32),
-  fontSizeHeading2: toRem(24),
-  fontSizeHeading3: toRem(20),
-  fontSizeHeading4: toRem(18),
-  fontSizeHeading5: toRem(16),
-  fontSizeHeading6: toRem(14),
-  fontWeightRegular: '400',
-  fontWeightMedium: '500',
-  fontWeightSemibold: '600',
-  fontWeightBold: '700',
-  // Transitions
-  transitionDuration: '200ms',
-  // Shadows
-  shadowResting: '0px 1px 2px 0px rgba(10, 13, 18, 0.05)',
-  shadowTopmost: '0px 4px 6px 0px rgba(28, 28, 28, 0.05), 0px 10px 15px 0px rgba(28, 28, 28, 0.10)',
-  // Focus
-  focusRingColor: colors.gray[1000],
-  focusRingWidth: '2px',
-}
-
-type NewTheme = typeof defaultTheme
 
 export const ThemeProvider: React.FC<ThemeProviderProps> = ({
   theme: partnerTheme = {},
@@ -110,15 +25,60 @@ export const ThemeProvider: React.FC<ThemeProviderProps> = ({
   const containerRef = useRef<HTMLElement>(null)
 
   useEffect(() => {
-    /**
-     * Merging partner overrides into default theme and injecting flattened css variables into document(scoped to .GSDK)
-     */
+    // Temporarily supports the legacy theme while we migrate to the new theme. Pulls out the
+    // legacy theme entries and merges them with the partner overrides. Once the theming migration
+    // is complete, we can remove the legacy theme entries and only support the new theme.
+    const {
+      colors,
+      spacing,
+      typography,
+      radius,
+      transitionDuration,
+      rootFS,
+      focus,
+      shadow,
+      input,
+      button,
+      radio,
+      checkbox,
+      table,
+      calendarPreview,
+      card,
+      link,
+      badge,
+      ...gustoSDKPartnerTheme
+    } = partnerTheme
+
+    const legacyPartnerTheme = {
+      ...(colors ? { colors } : {}),
+      ...(spacing ? { spacing } : {}),
+      ...(typography ? { typography } : {}),
+      ...(radius ? { radius } : {}),
+      ...(transitionDuration ? { transitionDuration } : {}),
+      ...(rootFS ? { rootFS } : {}),
+      ...(focus ? { focus } : {}),
+      ...(shadow ? { shadow } : {}),
+      ...(input ? { input } : {}),
+      ...(button ? { button } : {}),
+      ...(radio ? { radio } : {}),
+      ...(checkbox ? { checkbox } : {}),
+      ...(table ? { table } : {}),
+      ...(calendarPreview ? { calendarPreview } : {}),
+      ...(card ? { card } : {}),
+      ...(link ? { link } : {}),
+      ...(badge ? { badge } : {}),
+    }
+
+    const legacyTheme = createTheme(legacyPartnerTheme)
+
+    const gustoSDKThemeWithOverrides = {
+      ...gustoSDKTheme,
+      ...gustoSDKPartnerTheme,
+    }
+
     const theme = {
-      ...createTheme(partnerTheme),
-      /**
-       * Adding a string from translations for indicating optional form elements with CSS
-       */
-      optionalLabel: partnerTheme.optionalLabel ?? `'${t('optionalLabel')}'`,
+      ...legacyTheme,
+      ...gustoSDKThemeWithOverrides,
     }
 
     if (GThemeVariables.current) {
