+++ src/contexts/ThemeProvider/theme.ts
@@ -1,0 +1,93 @@
+import '@/styles/sdk.scss'
+import { getRootFontSize, toRem } from '@/helpers/rem'
+
+// Colors are for internal use in our theme currently
+// We don't export them for partner use or overrides
+const colors = {
+  neutral: {
+    100: '#FFFFFF',
+    200: '#FBFAFA',
+    300: '#F4F4F3',
+    400: '#EAEAEA',
+    500: '#DCDCDC',
+    600: '#BABABC',
+    700: '#919197',
+    800: '#6C6C72',
+    900: '#525257',
+    1000: '#1C1C1C',
+  },
+  error: {
+    100: '#FFF7F5',
+    500: '#D5351F',
+    800: '#B41D08',
+  },
+  warning: {
+    100: '#FFFAF2',
+    500: '#E9B550',
+    700: '#B88023',
+    800: '#B88023',
+  },
+  success: {
+    100: '#F3FAFB',
+    400: '#2BABAD',
+    500: '#0A8080',
+    800: '#005961',
+  },
+  info: {
+    100: '#F3FAFB',
+    400: '#2BABAD',
+    500: '#0A8080',
+    800: '#005961',
+  },
+  orange: {
+    800: '#CA464A',
+  },
+}
+
+export const gustoSDKTheme = {
+  // Colors
+  colorBody: colors.neutral[100],
+  colorBodyContent: colors.neutral[1000],
+  colorPrimary: colors.neutral[1000],
+  colorPrimaryContent: colors.neutral[100],
+  colorSecondary: colors.neutral[100],
+  colorSecondaryContent: colors.neutral[1000],
+  colorInfo: colors.info[800],
+  colorInfoContent: colors.info[100],
+  colorWarning: colors.warning[800],
+  colorWarningContent: colors.warning[100],
+  colorError: colors.error[800],
+  colorErrorContent: colors.error[100],
+  colorSuccess: colors.success[800],
+  colorSuccessContent: colors.success[100],
+  // Radius
+  buttonRadius: '6px',
+  inputRadius: '6px',
+  // Font
+  fontSizeRoot: getRootFontSize(),
+  fontFamily: 'Geist',
+  fontLineHeight: toRem(24),
+  fontSizeSmall: toRem(14),
+  fontSizeRegular: toRem(16),
+  fontSizeMedium: toRem(18),
+  fontSizeHeading1: toRem(32),
+  fontSizeHeading2: toRem(24),
+  fontSizeHeading3: toRem(20),
+  fontSizeHeading4: toRem(18),
+  fontSizeHeading5: toRem(16),
+  fontSizeHeading6: toRem(14),
+  fontWeightRegular: '400',
+  fontWeightMedium: '500',
+  fontWeightSemibold: '600',
+  fontWeightBold: '700',
+  // Transitions
+  transitionDuration: '200ms',
+  // Shadows
+  shadowResting: '0px 1px 2px 0px rgba(10, 13, 18, 0.05)',
+  shadowTopmost: '0px 4px 6px 0px rgba(28, 28, 28, 0.05), 0px 10px 15px 0px rgba(28, 28, 28, 0.10)',
+  // Focus
+  focusRingColor: colors.neutral[1000],
+  focusRingWidth: '2px',
+}
+
+export type GustoSDKTheme = typeof gustoSDKTheme
